podSearch-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yt-dlp==2023.12.30
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.32.3
defusedxml==0.7.1
supadata==1.2.1
python-dotenv==1.0.0

# MongoDB dependencies for efficient transcript storage
pymongo==4.6.1
motor==3.3.2
beanie==1.24.0
